Binary classification cluster bug vs. non-bugs
For each cluster of  bugs I built a binary classificator that distinguishes between that type of bug and non-bugs.
The model built was trained on versions 1.0.0-1.14.0 of the Calcite software and tested on version 1.15.0.
For each model, a report was generated, alongside a confusion matrix and a visual representation of the learning curve.